Rename section view to reader view
Update mixins to not depend strongly on eachother.
